don't test image manifest pulls on windows
